initial port to Windows
